feat(debug-log): create foundation of Roku debug log command
A Roku device has a debug log which is accessible via telnet.  We can use this to provide data via the WD `/session/:sessionId/log` route.

What's here so far is an implementation of a thing which just listens on an arbitrary telnet server and pipes its output to a file, since that's probably unwise to keep in memory.  When the command is issued, the driver can send the contents of the file to the client (and probably truncate it?).

This depends on type changes not yet published in Appium and will fail CI.

Note: The reason for two separate Telnet libs is that one provides a server & client and another only provides a client, but the client in the former is broken and I couldn't find another decent module which creates a dummy telnet server for testing.

- added wallaby config
- replace `chai` with `unexpected`